[3.3.1.3] * Opgelost een fout in het kiezen van een column kleur. * Toegevoegd de mogelijkheid om filter criteria's te vergelijken met een andere variable.  Het is daarmee mogelijk om te filteren of kleuren te laten verscheinen om basis van andere variable waardes. * Toegevoegd een eventuele 'Opmerking' in de hoofd Productiestatus Paneel. * Nog paar andere verbeteringen.

+++ Rpm/Productie/ProductieFormulier.cs
@@ -1809,7 +1809,7 @@ namespace Rpm.Productie
             var rooster = Manager.Opties.GetWerkRooster();
             if (aantal <= 0)
                 return Werktijd.EerstVolgendeWerkdag(DateTime.Now, ref rooster,rooster,null);
-            var peruur = (int) ActueelProductenPerUur();
+            var peruur = ActueelProductenPerUur();
             if (peruur == 0)
                 peruur = PerUur;
             var tijd = peruur > 0 ? aantal / peruur : DoorloopTijd;
